+++ src/router.js
@@ -12,7 +12,9 @@ import Payment from "./views/Payment.vue";
 import Stuff from "./views/Stuff.vue";
 import Admin from "./views/Admin.vue";
 
-import MenuItem from "./components/Admin/MenuItem.vue";
+import ChangeMenu from "./components/Admin/ChangeMenu.vue";
+import NewItem from "./components/Admin/NewItem.vue";
+
 
 import NotFound from "./views/NotFound.vue";
 
@@ -55,12 +57,17 @@ export default new Router({
     {
       path: "/admin",
       name: "admin",
-      component: Admin
+      component: Admin,
     },
     {
       path: "/change/:id",
       name: "change",
-      component: MenuItem
+      component: ChangeMenu
+    },
+    {
+      path: "/add",
+      name: "add",
+      component: NewItem
     },
     {
       path: '*',