+++ CHANGELOG.md
@@ -11,7 +11,7 @@ All notable changes to this project will be documented in this file.
 - Cart page
 - Menu details page
 - Component 'Navigation'
-- Componeent 'SummaryOrder'
+- Component 'SummaryOrder'
 
 ## [0.0.2] - 2019-02-08
 ### Added